pyproject.toml
[tool.poetry]
name = "tms-drfproject"
version = "0.1.0"
description = ""
authors = ["ig-rudenko <ig.rudenko1@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djoser = "^2.2.3"
pillow = "^11.0.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
drf-spectacular = "^0.27.2"
django-filter = "^24.3"
django-debug-toolbar = "^4.4.6"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
